Provide dynamic updated test metadata to main TFB results website
This change, along with some changes for the TFB results website, will
fix the problem of new tests being hidden until we update the metadata.
We'll still have to update the metadata for official rounds and previews,
but we won't have to for these continuous results.

Fetching the metadata is only possible once we have the results.zip file,
so I've hidden the "visualize this run on the TFB website" links until we
have the zip.  In other words, it is no longer possible to visualize an
in-progress run.