Fix visibility hacks that are now pointless because of nestmates
We avoided "private" in cases where another class in the same source file
(such as a nested class) would need access, and in order to grant this
access the Java compiler would have generated a synthetic accessor.  It
was a bad idea to have this performance / byte-code-size
micro-optimization in this codebase to begin with.  Now it's not an
optimization at all because Java 11 added "nestmates" and handles
same-source-file private access properly.