Use the JDK HTTP client instead of the Jersey client
There is an indentation change because the Jersey response objects used
try-with-resources blocks while the JDK response objects are not
closeable.  (The JDK response *bodies* are sometimes closeable, but in
the case of String or byte[] bodies, they are not.)

Because the JDK HTTP client supports HTTP/2 (and the Jersey client did
not), this migration exposed a bug in Undertow's server-side HTTP/2
handling of HEAD requests.  I've disabled HTTP/2 on the server.  See the
HttpServer class for details.