Show how to build and run the application with Docker
This is probably still not the recommended way of doing development on
this project, but it does work.  The Docker route may be more inviting
to people who don't routinely do Java development or don't have the
latest version of Java installed.  Even if no one runs the Dockerfile,
it's still theoretically useful as documentation.  "Here's what software
you need installed, here's how to build the app, here's how to run it."