Remove custom temp directory, use system-provided temp directory
This was unnecessary complexity.  The purpose of using an injected file
system instead of the default file system was so that, during tests, we
could write to the file system without having to worry about cleaning up
afterwards.  The same is already true with temp files, so it's ok if the
temp files remain on the real disk instead of the fake in-memory disk
during tests.