Update dependencies
Use maven.compiler.release instead of source+target, which corresponds to
the --release flag added in Java 9 to replace the source and target flags.

Upgrade to hk2 2.5.0.  This release is the first one after the project
tranferred ownership from Oracle to Eclipse, and so it comes with some
rebranding of dependencies (javax -> jakarta).

Remove <encoding>UTF-8</encoding> from the compiler settings, since the
default value for that is defined by project.build.sourceEncoding, which
we've set to UTF-8.