Optimize finding the file to replace when handling an upload
Before, we would parse every existing results file of the same type to
find the one with the same uuid.  We had the `UuidOnly` optimization to
make that parsing 2/3 as expensive as parsing a full `Results` object,
but it was still slow.

Now, we make use of HomeResultsReader's cache, which already associates
uuids with file names.

Since we're no longer calling `tryReadUuid` for a large number of files
within the scope of each request, there is much less benefit to
micro-optimizing `tryReadUuid`.  The only reason `UuidOnly` exists is to
micro-optimize `tryReadUuid`, so we can get rid of `UuidOnly` now.