Decouple run summary properties from specific file types
This lets us display runs on the home page and results detail page
properly when there is only a results.zip file and no standalone
results.json file.

Fixes #11

This also moves us closer to a fix for #12, since the properties that are
relative to the current wall clock time are no longer being cached.