Use nullness annotations from checkerframework instead of jsr305
jsr305 is problematic because it lives in the javax.annotation package,
which is shared by many other libraries, causing split package errors
when trying to do anything with Java modules.

The checkerframework @Nullable is also slightly superior to jsr305's in
that it is a type use annotation, as in `List<@Nullable String>`.  The
position of the @Nullable annotations has been adjusted to reflect this.

I had been holding back on this change because IntelliJ did not recognize
checkerframework's alternative to jsr305's @TypeQualifierDefault.  That
issue was fixed earlier this year.  checkerframework's alternative,
@DefaultQualifer, is not quite a drop-in replacement for us because I
couldn't figure out how to apply that to a custom annotation like our
@NonNullByDefault and have our custom annotation carry through the
nullness assumptions to the rest of the code.  Instead, I got rid of the
custom annotation and applied @DefaultQualifer directly to all packages.

The syntax for nullable arrays is unfortunate: `Foo@Nullable[]` would
indicate a `Foo[]` that may be null, whereas `@Nullable Foo[]` would
indicate an array of nullable `Foo`s.