Return to supported JDK version on non-alpine
There is no supported JDK 13 on alpine.  We were accidentally using a
pre-release version of JDK 13.  Use debian:buster-slim at runtime now
instead.

Add some commented-out logic to run tests using our modified JRE.  It
almost works, except that it skips a bunch of tests for whatever reason.
It would be nice to be in a situation where, if we forget to include some
module in our JRE that we need, our tests catch that mistake.

Add and use a "skipCopyDependencies" option to avoid copying our
dependencies into the lib/ directory twice during a full build.  Now, the
build can skip that step when the dependencies haven't changed.