Add tests for uploading results
It's a relief to finally have this.  The results upload handler may be
the most complicated part of the application and it's difficult to test
manually.  It's the piece I was always most worried about breaking
whenever I made code changes.

Add JSON deserialization annotations to a few view classes.  This allows
the test to examine the detail pages for the results it uploads.

In HomeUpdatesHandlerTest, add a `webSocket.request(1)` just so I don't
forget to add that the next time I use web sockets in a test.  It turns
out that line is necessary if you ever want to receive more than one
message over the socket.