Use dependency injection to reduce test-level setup boilerplate
Now if a test wants to refer to a service, it can add that service as a
method parameter, and the test infrastructure will inject an instance of
that service appropriately.

Since the injection behavior is centralized, it would be easy to change
the tests to share a single, global TestServices instance, which may
substantially reduce the total test time.  I haven't made that change,
but I did add a TODO item to consider it.

Remove a low-value assertion from HomeResultsReaderTest that implicitly
asserted that UploadResultsHandlerTest hadn't already been executed using
the same shared TestServices.

Add several TODO items to TestServices that would help us stop using the
TestServices class in tests directly.  Using TestServices directly is a
bit of a code smell in the same way that using a ServiceLocator or
another god object in production code would be; it shouldn't be necessary
when we have dependency injection and a sensible division of
responsibilities between classes.

Add TODO items for removing some test setup from @BeforeAll methods.
With the parameters being injected it's less clear that the @BeforeAll
methods are referring to the same instances of services that the @Test
methods are.  (Both currently *are* referring to the same instances, but
a sensible refactoring of TestServicesInjector could change that, so the
tests shouldn't make that assumption.)