Add tests for RootHandler
Modify ExceptionLoggingHandler so that it logs the exception before
committing the response instead of after.  This makes the behavior easier
to test because we're guaranteed that once we have the response, the
exception has been logged.

Improve RootHandler's documentation to discuss the features it provides
besides routing, which are the features we're testing.  All of the other
handlers' tests already verify that routing works.