Make various improvements to the Services class
Support nullable injected parameters in tests, removing an assymetry
between test parameter injection and hk'2 built-in parameter injection.

Remove (or make private) some methods of Services that were not being
called.

Remove the check that disabled per-lookup disposable services, because we
technically do expose an API for disposing of them.  Add tests that
exercise this kind of disposal.

Improve the naming and documentation of methods in the Services class.