Log exceptions after committing the response, reverting a previous change
It turns out that access logging occurs after the response, so the
justification we used for logging exceptions beforehand -- that it was
easier to test -- doesn't seem so compelling anymore.  Fix both logging
tests in RootHandlerTest to sleep before inspecting the logs.  Seeing the
access log test fail (because of inspecting the logs too early) is what
motivated this change.