Disentagle config deserialization logic from primary constructors
The primary constructors simply assign parameter values to fields of the
exact same name and type.  These constructors could be auto-generated by
various tools or, in a future verison of Java, be auto-generated record
constructors.  Default value subsitution is moved out of those
constructors and into the deserialization methods, since that's when we
want the default value substitution to occur.

Remove useless "root" property from AssetsConfig and MustacheConfig, which
we never want to change from its default value.