Support publish-subscribe communication between services
This should allow for much nicer separation of code that triggers an
event, such results being uploaded, from code that runs in response to
that event, such web socket listeners on the home page being notified,
and run progress timeouts being reset, and end-of-run emails being sent.
More specifically, UploadResultsHandler should not need to directly call
the HomeUpdatesHandler or RunProgressMonitor or EmailSender.  Instead, it
should publish an event that results have been updated, and then other
services who subscribe to those events should react accordingly.

The hk2-extras package provides a default implementation of
TopicDistributionService that was *almost* what we wanted, but which had
the fatal flaw that messages will only reach subscribers that have already
been created for other reasons.  That is, the delivery of a message won't
cause the subscribers to be initialized if they weren't already, and so
unless something else had a dependency on the subscriber and that
something else was initialized, the message intended for the subscriber
would be lost.

Aside from that major difference from hk2-extras, our homemade topic
distribution service tries very hard to behave just like the default
implementation and to implement all features correctly.

The current topic distribution service implementation runs on the same
thread as the caller.  Whenever a publisher publishes a message, that
publisher has to wait while all subscribers process that message on the
current thread.  We may revisit this later.  We want to use TaskScheduler
for running things async, but it would be bad for this core topic
distribution service to have a dependency on TaskScheduler.  HK2's
documentation shows a plausible workaround; there is an example where
they inject a secondary topic distribution service that delegates to the
first, where the second simply starts a new thread and calls the first
from that thread.  An approach like that could be fine, since it wouldn't
create weird tangles in our dependency graph for that second distribution
service to have a dependency on TaskScheduler.

We also looked at a couple of other publish-subscribe frameworks: Java's
built-in java.utill.concurrent.Flow, and Guava's EventBus.  Flow was far
too verbose and difficult to use.  EventBus was much better, except that
like the default hk2-extras implementation, it also had no great answer
for when/how to initialize subscribers.  We want subscribers to be
lazy-initialized in response to messages that they would receive, but
this is not possible with EventBus -- we'd have to eagerly initialize all
subscribers somehow.  Another benefit of HK's Topics over other
frameworks is the ability to inject other services into the subscribe
methods, but this is minor compared to the lazy-initialization issue.