Create a new dedicated service for sending the "run complete" emails
The new service subscribes to the same "results updated" events used in
other pub-sub functionality.  This caused an interesting side effect in
tests where, since another test was also triggering the "run complete"
email for whatever reason (perhaps it was the run progress monitor's test,
which adds a new results.zip file to disk), that meant our "don't send
these emails too quickly" guard kicked in and squelched the second email.

To resolve that test failure, I added a new configuration setting for
adjusting the time it waits between emails and set it to zero during
tests.

There sure is a lot of boilerplate involved in adding that one new
configuration setting.