Add support for @Provides annotations
This allows services to provide additional related services.

The most obvious use case for this on our application is on the
ApplicationConfig class, where we could annotate each of its fields with
@Provides instead of having a *ConfigFactory class for each one.

The Services class is becoming very large.  In later commits it will
probably be moved into the new hk2.extensions package and split apart into
mulitple top-level classes.