Use @Provides annotations to replace most of the *ConfigFactory classes
Unwrap the EmailConfig from Optional.  We've been dodging issues caused
by HK2's built-in support for Optional -- it allows all services to be
wrapped in Optional, but the feature seems half-baked and tends to throw
exceptions in cases where non-Optional injection would not.  The only
reason we avoided @Nullable before was that HK2 doesn't support singleton
nulls.  The new @Provides annotation will automatically use per-lookup
scope for @Nullable elements, so we don't have to think about that issue.