Move the Services class into the hk2.extensions package
Move its nested classes to the top level, since the file was too large.

Fixed a bug with PostConstruct methods not being invoked for static
@Provides methods and fields.

Added some more documentation.  Removed mentions of tfb-status-specific
stuff from the Services class, namely "HTTP handlers" and the HttpServer
class.

Moving all this to another package uncovered an interesting issue with
visiblity/accessibility.  It threw a new exception when trying to invoke
the @Provides method of ApplicationConfigFactory, because although the
method was public, the class was package-private.  We happened to avoid
this exception before because Services was in the same package as
ApplicationConfigFactory.  Now, this code will call setAccessible(true)
on the member it's trying to invoke when it's not already accessible.
This seems to be the exiting behavior of HK2 when it comes to everything
else, so we might as well follow suit.