Refactor some classes to use @Provides instead of an HK2 Factory
Only InMemoryFileSystemFactory is still a Factory because there is
currently no equivalent of factory.dispose(...) for services that come
from @Provides.

Keep the "Factory" naming scheme of these classes for now, but that is
likely to change later.  (We probably wouldn't have used the "Factory"
suffix in the first place if we'd had @Provides all along.)