Do some renaming, adjust static field handling, improve documentation
Rename @Provides(destroyMethod = ...) to @Provides(disposeMethod = ...).
Rename related fields and types accordingly.

Do not attempt to dispose of instances that come from static fields.
The lifetime of such an instance is tied to the lifetime of the class,
which probably outlives the ServiceLocator.  Presumably, disposing of
that instance would leave it in an unusable state, making the class
itself unusable from that point forward.  It is probably a bad idea in
general to store anything that requires disposal in a static field.
Adjust the disposal tests for static fields accordingly.

Show example usage of @Provides in documentation.

Remove special handling for Factory classes, which introduced a dependency
on the internal Utilities class from hk2.  This change means that the
Factory classes themselves will be scanned for @Provides annotations
instead of the classes they return from their provide() methods.  It's
unclear if the previous behavior was better, but since there is no reason
to use a Factory when @Provides is available anyway, supporting Factory
seems like a non-priority.

Try to improve naming of internal methods, parameters, and so on.