Do more cleanup of results testing code and sharing
Include the share id in ShareResultsJsonView so consumers don't need to
parse it out of other fields.  Remove the JSON file name field, which is
no longer used.

Add test library methods for viewing Results as InputStream, ByteSource,
and HttpRequest.BodyPublisher.  This lets us remove a bunch of code that
was writing the Results to temporary files.

Refactor multipart/form-data submission code to stream the files instead
of buffering them all up front in memory before sending the request.