Visualize in-progress runs, prepare for cached_query rename
Historically we could not visualize in-progress runs because we didn't
have the test metadata required by the TFB website.  Finished runs had
them in the form of a test_metadata.json file in the results.zip file.
Now, the toolset has been updated such that results.json files embed
their own test metadata, and the TFB website has been updated to use this
embedded test metadata when available.  Therefore, we can start rendering
the "visualize" link for a run as soon as we see test metadata, which
might be right away.

Prepare for the "cached_query" test type to be renamed to "cached-query".
We already have lots of historical data using the "cached_query" name,
and we don't know when the toolset will be changed to start using the new
name, so we'll support both names.  Supporting multiple names makes it
more appealing to use enum constants instead of strings to represent the
test types, which is a change we already wanted to make for other reasons
(the test types have always been enum-like), because this lets us confine
the test-rename-aware logic to the serialization layer and use the enum
constants everywhere else.