Remove the /attributes and /saveAttributes endpoints
These endpoints had two audiences:

1. /attributes was called by the TFB website when rendering results by
   run id for a continuous run.  The TFB website needed this to obtain
   the test metadata for that run.  This endpoint also did the work of
   transforming the raw test metadata into a minified,
   TFB-website-specific format.  It was probably a mistake to ever have
   tfb-status do this conversion instead of keeping that logic in the TFB
   website.
2. /saveAttributes was theoretically called by an admin user for each
   official round, uploading the newest set of minified metadata to use
   as a baseline for continuous runs.  As far as I know, this never
   occurred in practice.  The last modified time of the file in production
   confirms this.  Since we've apparantly gotten by without ever using
   this endpoint, it's likely the whole procedure we imagined around it
   was unecessary.

Now, (1) is gone because the TFB website never calls this.  It either uses
the raw test metadata embedded in the results.json, or it calls the
/unzip endpoint of tfb-status to extract the test_metadata.json file if
that's available.  And now, the TFB website does that transformation of
the raw test metadata into that minified rendering format itself.  (2) was
never a real audience, only imagined, so that's gone.

This code being deleted was useful while writing its replacement, but it
was difficult to understand and maintain, and it was untested.  It will
not be missed.