Support routing by request media type
Attach the matched @Route annotation to each request.
UploadResultsHandler demonstrates a use for this:  it declares multiple
routes, and it inspects this attachment to see which one was matched.

Remove support for method="*" routes, which was only used by TestHandler.

Fix wrong exception message in MethodHandler.addMethod.  Add shortcut for
adding a handler with a String method instead of an HttpString.