Support routing by the "Accept" header
Add the AcceptHandler class, which implements the core logic for routing
to different handlers depending on the Accept header.

Add a `produces` field to the @Route annotation, where the value of this
field is input to an AcceptHandler.  Use this field in HttpServer,
implementing its advertised behavior.

Since using the `produces` field will set the Content-Type header of
responses automatically, prefer that to manually setting the Content-Type
of the response within the handleRequest method.

Prefer application/json as a response type over JSON_UTF_8, which
redundantly specifies the utf-8 charset even though that is implied by
application/json.  This media type is being used in @Route annotations,
which can't refer to JSON_UTF_8, so there's no longer as much "use the
predefined constant even though it's not exactly right" pressure.

Complete TODOs in MediaTypeHandler for using a better algorithm for
media type specificity.  This algorithm was required by AcceptHandler, so
it was added to a new, package-private MediaTypes utility class.

Improve MediaTypeHandler's documentation to be more specific about how it
matches request media types to handlers, since it and AcceptHandler use
different matching algorithms.