Update dependencies
Hold back on updates that would migrate from javax.* to jakarta.* package
names.  hk2 3.x is available, but that new version updates many of hk2's
dependencies with javax->jakarta renames, and it's not clear that the rest
of our dependencies are compatible with that.

Update the documented command for checking dependency updates.  The new
command finds plugin updates missed by the old command.

Ignore a different spelling of Jackson release candidate versions.