Migrate from javax. to jakarta. packages
Use alpha version of greenmail.  The current stable version is tied to
javax.mail, which holds everything else back.  The mail package has a
hidden dependency on inject in the form of a META-INF/hk2-locator/default
file, causing errors if we don't upgrade mail and inject in sync.  This
was preventing us from upgrading hk2, which now uses jakarta.inject.