Use angus-mail and angus-activation, upgrade greenmail
The situation with the mail and activation packages is a mess due to all
the renames and changes of ownership.  I think I've got a handle on it
now, though.  The only problematic package is
org.eclipse.angus:jakarta.mail, which seems to duplicate classes that are
contained in its own dependencies.  I assume that was a mistake in their
release process.  Thankfully, everything seems to work out if we exclude
that dependency.